requirements.txt
# requirements.txt

aiosqlite==0.21.0
alembic==1.14.1
anyio==4.8.0
APScheduler==3.11.0
asyncpg==0.30.0
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-telegram-bot==21.10
pytz==2025.1
requests==2.32.3
sentry-sdk==2.22.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0